+++ src/robodash/views/selector.cpp
@@ -78,7 +78,9 @@ void rd::Selector::sd_load() {
 	}
 
 	for (rd::Selector::routine_t &r : routines) {
-		if (strcmp(r.name.c_str(), saved_name) == 0) selected_routine = &r;
+		if (strcmp(r.name.c_str(), saved_name) != 0) continue;
+		selected_routine = &r;
+		run_callbacks();
 	}
 
 	if (selected_routine != nullptr) {
@@ -105,6 +107,8 @@ void rd::Selector::select_cb(lv_event_t *event) {
 
 	selector->selected_routine = routine;
 	selector->sd_save();
+
+	selector->run_callbacks();
 
 	if (routine == nullptr) {
 		lv_label_set_text(selector->selected_label, "No routine\nselected");
@@ -208,9 +212,28 @@ rd::Selector::Selector(std::string name, std::vector<routine_t> new_routines) {
 
 // ============================= Other Methods ============================= //
 
+void rd::Selector::run_callbacks() {
+	for (select_action_t callback : this->select_callbacks) {
+		if (this->selected_routine == nullptr) {
+			callback(std::nullopt);
+		} else {
+			callback(*this->selected_routine);
+		}
+	}
+}
+
 void rd::Selector::run_auton() {
 	if (selected_routine == nullptr) return; // If commanded to do nothing then return
 	selected_routine->action();
+}
+
+std::optional<rd::Selector::routine_t> rd::Selector::get_auton() {
+	if (selected_routine == nullptr) return std::nullopt;
+	return *selected_routine;
+}
+
+void rd::Selector::on_select(rd::Selector::select_action_t callback) {
+	select_callbacks.push_back(callback);
 }
 
 void rd::Selector::focus() { rd_view_focus(this->view); }